Gamium-FinanceAI/requirements.txt
# Gamium Finance AI - Dependencies
# Python 3.9+

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
gunicorn
pyyaml
requests>=21.0.0

# Deep Learning
torch>=2.0.0
numpy>=1.24.0

# Reinforcement Learning
gymnasium>=0.29.0

# Data Processing
pandas>=2.0.0
scipy>=1.11.0
pyarrow>=14.0.0
fastparquet>=2023.0.0

# Visualization
matplotlib>=3.7.0
rich>=13.0.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
requests>=2.31.0  # 用于云存储下载
